First stone of the project
Set the project as a 3D unity project with the goal to create a tower defense game. So far, only a cube has been created, with a script called 'Enemy' with the Debug.log info on start. The goal is to make sure other members of the projects are able to download properly the project.